third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-prefetch-in-main-frame.html

<!DOCTYPE html>
<html>
<body>
<p>
Prefetching from insecure source must trigger a mixed content callback.
</p>

<link rel="prefetch" href="http://example.test:8080/resources/square.png"  onload="prefetch_onload()" />

<script>
if (location.protocol != 'https:')
    location = 'https://127.0.0.1:8443/security/mixedContent/insecure-prefetch-in-main-frame.html';

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function prefetch_onload() {
    if (window.testRunner)
        testRunner.notifyDone();
}
</script>
</body>
</html>
